backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.12
torch==2.5.1
numpy>=2.0.0
pandas>=2.2.3
mediapipe>=0.10.14
opencv-python-headless>=4.10.0
websockets==14.1
scikit-learn>=1.6.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
